No reason for putting the manifest in the ${basedir}
The one single reason for doing it is to allow to import in the workspace the
`sbt.full.library` project, so that it can be correctly consumed by the
dependent projects (e.g., the `core project). However, given we practically
never had to do so in the past 6 months, it feels reasonable to remove the
flexiblity in the name of a cleaner build. In case you will need to import he
`sbt.full.library` project in Eclipse at any point, just create the required
.classpath and .project, and copy/paste in the project's root the META-INF
folder generated by the build (look in the target/ folder). Make  sure to do
the same for the `sbt.compiler.interface` project.