Deprecated `doWithPresentationCompiler` and `withPresentationCompiler` in ScalaProject
ScalaProject's methods `doWithPresentationCompiler` and
`withPresentationCompiler` are superseeded by `presentationCompiler.apply`. The
main reason for doing so is that I don't think it is useful to have ARM methods
when the used resource is publicly accessible (i.e., the `presentationCompiler`
field).

While doing this refactoring, I noticed we need a 'null object' of a `Response`
(look at `ScalaSourceFile.scheduleReconcile()`).  The only alternative to
returning a 'null object' is to throw an exception, which I don't think is the
right solution in this case. Unfortunately, creating a 'null object' for
`Response` is currently cumbersome because `Response` is a concrete class, and
not an interface. Hence, I'm deferring this further refactoring to a different
pull request.