`retryFlakyTest` now enabled by default
As the last failures of the PR job validator demonstrate, it's quite easy to
get several flaky builds in a row. This is painful to a degree that cannot be
tolerated anymore, which is why I'm enabling the `retryFlakyTest` flag by
default, and you should explicitly pass `-DretryFlakyTests=false` to the build
if you don't want known flaky test to be re-run.

By the way, I'm not removing the `retryFlakyTests` profile because the scala
validation script uses it, and this way we can easily change again the default
of `retryFlakyTest` without having to go and update the validation script.