Added autocomplete for constructors.
This was done by adding an additional case to the completion logic
in ScalaCompletions.

To avoid problems with completion overwrites, similar to Eclipse Java
only tooltips are shown for constructor completion.

A known defect is the ability to complete constructors of classes with
mixins on the same line. See #1001950

Fix #1001272