Hook back the reconciling participant extension point.
Removed the old `reconcile` override, based on the Java reconciler (no longer
used), and moved the only useful piece of functionality in the existing
`reconcile` method: the calls to before/after reconciliation.

There might be a better way to handle reconciliation decorations, but that
requires more work, and this PR should be focused only on quiescence.