Remove trailing whitespace in field completion
A field in Scala consists a getter method and an underlying field that
has the same name as the getter. Because of this there would exist two
Symbols with the same name in the same scope, which is not allowed.
Thus, the underlying field contains an trailing whitespace to separate
it from the getter.

This fix just trims all Symbol names coming from scalac to avoid moving
the trailing whitespace around.

Fixes #1001973