Don't create a reconciler for non-editor source configurations.
In the Syntax Coloring preview pane we create a source viewer without an
editor. The reconciler needs an editor to get to the compilation unit, so
we don't attach a reconciler in this case (there won't be any editing).

We might need to revisit this if we ever want to use a source viewer outside
an editor (with reconciliation capabilities).