Implementation of an indent guide + test suite
The indent guide is enabled by default and can be disabled in
  Preferences -> Scala -> Editor

The implementation tries to redraw as less as possible to guarantee no
delay while typing - some short tests measured a drawing time of 1ms
while typing and 20-30ms when a new file is opened.

Furthermore, the editor preference page is redesigned.

Fixes #1002033