Combines toolbar menu structures
The "show inferred semicolons" action was placed in its own menu
structure, whereas other actions are combined under a different menu
structure. This commit merges both of them.

For this to happen the deprecated usage of an Eclipse action in the show
inferred semicolons action is replaced with the newer handler
architecture.

This also moves the command for toggle mark occurrences to a different
location in the pom.xml.

Furthermore, a preference to toggle the "show inferred semicolons"
action is added to the editor preference page.

This does not introduce functional changes to the logic of the
"show inferred semicolons" action.

Fixes #1002042