Replace old action API used by refactorings with newer handler API
The names of refactoring classes that provide an integration into the UI
of Eclipse have changed, in order to minimize overhead in the class
names.

The previous code contained some logic that should disable a menu item
when nothing is selected in the editor. But the existing logic was
wrong, it did never work. Because I did not find out how to implement it
correctly with Eclipse core-expressions I romved the logic completely.

Formerly, the context menu item, that showed the refactorings, contained
an item to the "refactoring qick menu", which has now been deleted. This
menu contains all available refactoring and it can be display by
entering a key combination. This item doesn't provide additional
information, thus the deletion. Users can always look up the key
combination in the preferences.

This also changes the ids of the refactring commands.

Fixes #1002049