Introduce a migration manager for key preferences/bindings
Whenever the command ids of key bindings, that are defined by users,
are changed all the preferences are lost because the new commands do no
longer consider their values. To avoid this problem the class
`MigrationPreferenceInitializer` is introduced which can copy old
preference values that are defined by users and set them as default
values for the new preferences.

This class also undefines the old command id and therefore the old key
binding after the migration is done.