Add auto indent on tab strategy
This component is needed because in a former commit the "indent on tab"
action of the Java editor has been disabled. This auto edit strategy
choses Scala preferences instead of the Java ones and is enabled
whenever the tab key is pressed.

If a user selects a text region or pressed CTRL+Tab, which decreases
the current indentation, logic of the text editor is invoked instead of
the new auto indent strategy. The text editor built-in logic is of good
enough quality, thus it is not necessary to replace it. Nevertheless,
it is not extensible, which is a problem because there are still some
cases where its behavior could be improved a little bit.

If the new line key is pressed the already existing but very erroneous
`ScalaAutoIndentStrategy` is invoked. It can't be replaced easily
because in order to insert the right indentation after a new line a
more accurate, token based, logic needs to be written, which is able to
analyse the token that surround the cursor.

This commit corrects some tests for the multi line string auto indent
strategy, whose indent logic is moved to the new indent strategy.

There is also a new preference added to the editor preferences, which
allows to disable the new indent logic. The result is that when a tab
key is pressed only a single indent step is inserted, which is either a
tab or some spaces - as specified by the Scala formatter preferences.

The only downside of the new logic I can see so far is that in a case
like

    class X {
      def f = {
    ^
      }
    }

the result is

    class X {
      def f = {
      ^
      }
    }

instead of

    class X {
      def f = {
        ^
      }
    }

which was the old and better behavior. But because one needs to hit the
tab key for this special case only two times instead of a single time,
the change shouldn't be too annoying. Having the old logic is difficult
to implement because the previous token have to be analyzed.

Even this small flaw can't deceive that the new logic is now completely
based on Scala preferences. The only exception of this is the
`ScalaAutoIndentStrategy`, which internally still reads the Java
preferences.