Removal of unnecessary adjusting of cursor position
It turned out that in the completion proposal logic the whole logic to
adjust the cursor position is unnecessary because it can be done
together with applying the text change.

Furthermore it turned out that the dependency to the selection provider
can be removed as well because the text viewer which is needed and
provided by the selection provider is already passed to the completion
proposal.