Move more completion test cases to new test suite
All test cases that belong together are moved to the new test suite.
This does not include ignored tests that test behavior which is not
implemented yet. It does also not include tests that mostly test a
different part of the completion behavior. It turned out that #1001014
can be closed because it is fixed.

While writing these tests the new test suite had to be extended further
times. Nevertheless, there are still tests that can still not be
implemented with the new test suite - therefore these tests aren't moved
either.

Re #1001014