Expression evaluator for Scala IDE
Implementation of expression evaluator and conditional breakpoints based
on JDI.

Allows user to evaluate any Scala expression in context of some breakpoint
during debug. Expression is transformed in multiple phases and evaluated
using calls to remote, debuged machine and result is fetched back to user.

Because of complexity of Scala language, implementation details of JVM
platform and necessity for manual evaluation not all Scala features are
supported yet.

Co-authored-by:
 - Pawel Batko <pawel.batko@gmail.com> (@pbatko)
 - Piotr Kukielka <piotr.kukielka@gmail.com> (@pkukielka)
 - Jerzy Muller <sidus.gemini@gmail.com> (@Kwestor)
 - Michal Pociecha <michal.pociecha@gmail.com> (@mpociecha)
 - Krzysztof Romanowski <romanowski.kr@gmail.com> (@romanowski)