Integrate extraction refactorings into quick assist.
Extends the quick assist menu with the identical extraction proposals as shown by the "Extract ..." action in the refactoring menu. The only difference from a users perspective is, that selections of partial trees are not expanded.

Furthermore, the generation of a locally unique identifier for the extracted abstraction has been moved to scala-refactoring (see https://github.com/scala-ide/scala-refactoring/pull/51).