Work around race condition between presentation compiler and indexing
It could be that the Java indexer interfered a test by restarting the
presentation compiler. This happened because the new text edit test
suite creates files at runtime. This on the one sided triggered a
restart of the presentation compiler and on the other side triggered the
Java indexer, which implicitly restarts the presentation compiler too.
The problem here is that the Java indexer is scheduled too slowly, which
led to a race condition when the restart happened at the wrong time.