The ScalaBuilder chooses the Scala compiler version based on the project class path.
This commit adds the ability to build with either the current (at the time of this
work, 2.11) Scala compiler, or a previous version. The previous version is chosen
when a project is in “compatibility” mode, and the `-Xsource:` flag is used to put
the type-checker in the right mood.

The Scala builder now chooses a ScalaInstallation based on the Scala version on the
class path. The builder infers what Scala installation best matches the Scala Library.
It works by using the resolved class path, meaning that the Scala class path
container (if present) is honored as-is. If the `-Xsource` flag is set, but the Scala library is
still the platform version (2.11 at this time), the Scala library *on the classpath*
what determines the builder version of Scala.