Add validation logic to `ScalaFileCreator`
There is one flaw left in this validation logic: It only checks if an
identifier is a valid Java identifier and if it is no Scala keyword. For
package paths it should also check if they are no Java identifiers and
for file names it should check if the name is a valid Scala identifier
(which can contain special signs).