Validating logic now checks Scala and Java keywords
Previously it was only checked if identifiers can not correspond with
Java identifiers which is not enough for Scala identifiers.

Furthermore package names are not allowed to contain Java identifiers.
The Eclipse package hierarchy view can't display them correctly and even
the Java compiler of JDT seems to have problems with them.