Moves "auto insert semicolon at correct position" feature to Scala editor
This feature always moves the semicolon to the end unless the cursor is
inside a for-comprehension - the only place in Scala where the usage of
semicolons is a syntax feature.

The "auto insert braces at corret position" is no longer supported by
the Scala editor because there are very few cases where this feature
really makes sense. Most of the time one can use a template to generate
code where the opening brace is at the correct position.

Fixes #1002056