Optionalize automatic close operations
The behavior of auto closing braces is complemented by another option.
Whenever braces should be closed automatically it can also be checked if
the close is necessary and if not then it does not happen. This behavior
already existed but the preference was difficult to understand, thus its
description changed.

All other characters that need a closing sign got a new preference. They
are enabled by default, with exception of angle brackets because in
Scala they are rarely used for surrounding purposes.

This also removes a hack in `SurroundSelectionStrategy` to prevent the
addition of a closing `>`.

Fixes #1001716