Add adapter to `INewWizard` for `NewFileWizard`
The `NewFileWizard` interface could not be integrated into the Eclipse
platform. The problem is that various components of Eclipse, namely the
quick access component, the new file menu and the "New" dialog, can only
load entries that implement the `INewWizard` interface. Therefore an
adapter had to be introduced that allows to open the `NewFileWizard`
through the `INewWizard` interface.