Add registry extension mechanism to codebase
This component is required because it is not always possible to
configure the behavior of the IDE just with XML files.

In the present case, wizard elements defined by the fileCreator extension
point must be added to the platform. Because the fileCreator extension
point is completely independent from the newWizard extension point,
there doesn't seem to be an easier way to hook the additional components
into the platform than by using reflection.

The new `RegistryExtender` component implements this extension behavior.
It is important that this component is called on startup of the Scala
IDE bundle, otherwise it may be loaded to late and some Eclipse
components have already cached some registry entries, which may be
changed by this `RegistryExtender`.