Update shown wizards in Scala perspective
All changes done to a `IPerspectiveFactory` can only be seen when a new
perspective is created, which means that the changes done by this commit
don't have any effects to users that already use the Scala IDE.

Eclipse doesn't seem to support a way to remove and recreate a
perspective when the corresponding bundle is still installed in the
Platform. Therefore users first have to remove the Scala IDE bundle,
then delete the Scala perspective if it is not automatically removed and
finally install the new Scala IDE bundle.