Update shown wizards in Scala perspective
Because the old wizard is going to be removed, there is no need anymore
to show it in the UI. All UI entries now direct to the new wizard.

This also cleansup the `PerspectiveFactory` and removes the
`NetProjectWizId` and the `LogView` because the components they belong
to seem to be gone/unused for a long time already.