Fix problem with incorrectly updated editor area in `NewFileWizard`
Previously the editor was opened and immediately after the opening was
triggered the editor contents were touched by the `NewFileWizard` to
insert the content of the template. Sometimes this lead to an
incorrectly updated editor area - the content was simply not correctly
displayed.
This is fixed by first touching the content and then opening the editor.