Change validation logic of `ScalaFileCreator`
Originally the `ScalaFileCreator` was created with the goal in mind to
create files outside of source folders. But because the Eclipse based
template creation logic couldn't handle this feature (it needs a source
folder to open an Scala editor with template engine support) only empty
files can be created. Because of this the validation logic needs to be
changed too, otherwise it is not very clear to users if the wizard now
creates an empty file or a Scala class if they forgot to prefix the file
name with a source folder.

This also removes some tests whose behavior is unsupported now.