Create better initial paths in `ScalaFileCreator`
The new inital paths always start with a source folder. This makes it
very easy to create a new file without the need to think about how
source folder is named.

This also removes obsolete code about the folder notation. The folder
notation is no longer supported. If one wants to create a file that does
not exist in the source folder one can use the `EmptyFileCreator` or the
default Eclipse new file wizard.