Enable only one new Scala wizard instead of all of them
The new file wizard is a new component and it is not yet clear if users
will like it. Therefore instead of removing the old wizards altogether
and adding only the new ones, only one new Scala file wizard is added to
the group of existing Scala wizards.

The implementation adds also one new file creator, which basically
forwards to the Scala class creator. The new file creator is filtered
out wherever it is found or replaced with the class creator. The only
way to call this new file creator is to search for it in the search
field with "New Scala File" or to navigate to the "New" menu.