Move source folder selection to project selection field
It turned out that it is not very useful to let users specify the source
folder of a file in the file name text field. Therefore they now have to
select a source folder in a separate field. Formerly they had to select
the project in the separate field which changed insofar that only it is
only valid to select source folders.