Make `FileCreator` and `NewFileWizard` project independent
One problem of the original design of the wizard was that it assumes
that all files that can be created must be part of a project. While in
Eclipse this is generally true, it is not very future proof to design
the wizard with this in mind. Especially because the `ScalaFileCreator`
also relies on the existence of a source folder, this design would mean
that it is not easily possible to store files outside of projects.

The new designs is based entirely on folders, which can return a project
if it exists, but don't need to. Furthermore this allows the wizard to
express source folders easily, without the need to encode it in a weird
way into the name of the file.

This redesign also fixed some weird bugs related to source folders that
are subfolders of other folders of a project. For example if there
exists a source folder `dir/subdir/src` it would have been required to
encode the outer folders as part of the file name in order to allow a
`FileCreator` to create correct paths.

A positive advantage of the redesign is that the implementation of
`ScalaFileCreator` could have been simplified because they no longer
need to decode the file names.