Enable accidentally disabled inferred semicolon painter
This is by far not the best fix, because it is working around of the
purpose of the `EditorPainter` superclass. This class was introduced to
easy the implementation of a painter, but as it turns out now its design
is not scalable.

In Eclipse one has to install a listener to a document to get notified
when a change happens. This model does not work together with the
`EditorPainter` class because it would have to install all listeners the
subclasses need and not forward the occurring events when the painter is
disabled.

It would be better if the event wouldn't be sent at all when the painter
is disabled, but that is more or less a design limitation of Eclipse,
which we can't easily work around.

For now complexity needs to be added again to the subclasses of the
`EditorPainter`, in this case to `InferredSemicolonPainter`.

Fixes #1002153