Don’t lose warnings when building incrementally.
The builder clears all problem markers in a  given project, but not
all warnings will be reported in an incremental compiler run. For example,
the change might not require recompilation of a file that contains deprecation
warnings. In this case we’ll add markers based on the Analysis object supplied by Sbt.


See https://github.com/sbt/sbt/issues/1372

Fixed #1002148