Simplify `SaveActionSetting`
The properties `textBefore` and `textAfter` are replacey by
`codeExample`. The reason for this is that it is more useful when the
IDE computes the changes a save action is doing instead of hardcoding an
example into the sources.

However, without the ability to create save actions at runtime, this
functionality can't be implemented yet because we can't inject
dependencies easily into the save action instance.