Make save action computation async but don't abort them
Async save actions allows to return after a given timeout. Because the
API of the save actions doesn't provide a way to cancel further
computations of the save action once it timed out, the save action must
complete.

If it does not it will consume CPU and memory and the best seems to be
to restart Eclipse.