Configure information control creator when no SCU is available
The method `getInformationPresenter` is called not only for full editors
but also for "static" ones, i.e. the editors that are configured in the
preference pages. Because there exists no ScalaCompilationUnit for the
latter, the information presenter couldn't be correctly configured.

Now, a default information control creator is used, which is only used
for the "static" editors. All other editors (the editors for which a SCU
is available) use the non-default implementation.