Create only one instance of ScalaDocumentProvider
We need a way to access existing Scala compilation units in several
parts of the IDE. Therefore the document provider is moved to the Scala
plugin in order to act as a compilation unit cache. The Java plugin
works similar and has with the WorkingCopyManager a class that provides
such a cache, but because this class is statically configured in the
Java plugin we can't override with our own implementation.

Therefore, all calls to

    JavaPlugin.getDefault.getWorkingCopyManager.getWorkingCopyManager

need to be replaced with

    ScalaPlugin.plugin.scalaCompilationUnit

We can do these change in our own codebase, but there is the risk that
code inherited from JDT also calls accesses the working copy manager and
in these cases we have to override such implementations in order to get
the code up and running again.