Add possibility to update cursor to auto edits
This adds a new `Change` object to pass an updated cursor position from
an auto edit to the logic that applies it.

This also adds some syntax sugar for auto edits to operate on the input.

The return type of the auto edit has changed to an `Option` because it
is possible that an auto edit can not be applied. Previously the input
change object had to be returned to describe this situation but
returning an `Option` makes more sense.