Interrupt Scala builder on resource change
Whenever Eclipse is configured to automatically build, an interrupt
event is sent to all builders whenever an resource change happens in the
workspace.

The Scala builder didn't react on such an interrupt event so far - this
changes with this commit.

Beside from aborting the current running build, the build is restarted
after the resource change event is handled.

Fixes #1002229